requirements.txt
yfinance
pandas
numpy
scikit-learn
google-cloud-bigquery